Add version 1.6.4 of AngularJS as debian/missing-sources/angular.js
Downloaded from
https://ajax.googleapis.com/ajax/libs/angularjs/1.6.4/angular.js

Fixes lintian-reported error "source-is-missing".

This complements tests/end2end/data/hints/angular1/angular.min.js
which is included in upstream's test suite (which itself is included
on request by Gentoo and Arch packagers since 1.12.0), but not in the
preferred form for modification, just minified.

Also add according file stanzas and the text of the MIT license to
debian/copyright.

Additionally override source-contains-prebuilt-javascript-object as
upstream declared this inclusion of minified JavaScript code as being
intentional.